+++ src/Node.h
@@ -13,17 +13,17 @@
 class Node {
 
 	int id;
-	int lat;
-	int longe;
+	double lat;
+	double longe;
 
 public:
 	Node(){}
-	Node(int id, int latitude, int longitude);
+	Node(int id, double latitude, double longitude);
 
 
 	int getID() const;
-	int getLatitude() const;
-	int getLongitude() const;
+	double getLatitude() const;
+	double getLongitude() const;
 };
 
 bool operator==(const Node& n1, const Node& n2);
